requirements.txt
flask==2.0.1
werkzeug==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
markdown==3.3.4